requirements.txt
asgiref==3.3.4
dj-places==4.0.0
Django==3.2.3
django-countries==7.2.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
Pillow==8.2.0
psycopg2==2.8.6
PyJWT==2.1.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
